requirements.txt
pandas==2.2.2
numpy==1.26.4
openai==1.40.6
httpx==0.27.0
python-dotenv==1.0.1